refactor: storage as source of truth, local mortgage only, settings and result UI
