--- src/pages/MortageResult/index.ts
@@ -1,1 +1,0 @@
-export { default as MortageResult } from './MortageResult';
