+++ src/pages/MortgageResult/index.ts
@@ -1,0 +1,1 @@
+export { default as MortgageResult } from './MortgageResult';
