+++ src/providers/MortgageProvider.tsx
@@ -1,86 +1,21 @@
-import React, { createContext, useContext, useEffect, useState } from 'react';
-import type { LoanDetailsValues, EarlyPayment, RegularPayment } from '@/domain';
-import {
-  mortgageService,
-  MortgageCalculationResults,
-  AmortizationScheduleResults,
-} from '@/services/mortgage';
+import React, { createContext, useContext } from 'react';
+import type { CalculationWithResultsValue } from '@/hooks/useCalculationWithResults';
 
-interface MortgageContextType {
-  loanDetails: LoanDetailsValues | null;
-  setLoanDetails: (values: LoanDetailsValues) => void;
-  earlyPayments: EarlyPayment[];
-  setEarlyPayments: (payments: EarlyPayment[]) => void;
-  regularPayments: RegularPayment[];
-  setRegularPayments: (payments: RegularPayment[]) => void;
-  mortgageResults: MortgageCalculationResults | null;
-  amortizationResult: AmortizationScheduleResults | null;
-}
+export type MortgageContextType = CalculationWithResultsValue;
 
 const MortgageContext = createContext<MortgageContextType | undefined>(undefined);
 
-export function MortgageProvider({ children }: { children: React.ReactNode }) {
-  const [loanDetails, setLoanDetails] = useState<LoanDetailsValues | null>(null);
-  const [earlyPayments, setEarlyPayments] = useState<EarlyPayment[]>([]);
-  const [regularPayments, setRegularPayments] = useState<RegularPayment[]>([]);
-  const [mortgageResults, setMortgageResults] = useState<MortgageCalculationResults | null>(null);
-  const [amortizationResult, setAmortizationResult] = useState<AmortizationScheduleResults | null>(null);
-
-  // Single effect: base summary (loan only) + schedule (loan + overpayments).
-  // Ignore results if inputs change before promises resolve (avoid stale state).
-  useEffect(() => {
-    if (!loanDetails) return;
-
-    let cancelled = false;
-
-    mortgageService
-      .calculateBase(loanDetails)
-      .then((results) => {
-        if (!cancelled) setMortgageResults(results);
-      })
-      .catch((err) => {
-        if (!cancelled) console.error('Error calculating mortgage results:', err);
-      });
-
-    mortgageService
-      .generateAmortizationSchedule({
-        ...loanDetails,
-        earlyPayments,
-        regularPayments,
-      })
-      .then((results) => {
-        if (!cancelled) setAmortizationResult(results);
-      })
-      .catch((err) => {
-        if (!cancelled)
-          console.error('Error generating amortization schedule:', err);
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [loanDetails, earlyPayments, regularPayments]);
-
-  const value = React.useMemo(
-    () => ({
-      loanDetails,
-      setLoanDetails,
-      earlyPayments,
-      setEarlyPayments,
-      regularPayments,
-      setRegularPayments,
-      mortgageResults,
-      amortizationResult,
-    }),
-    [
-      loanDetails,
-      earlyPayments,
-      regularPayments,
-      mortgageResults,
-      amortizationResult,
-    ]
-  );
-
+/**
+ * Provides mortgage calculation state to the result page subtree.
+ * Only wrap result page content when you have value from useCalculationWithResults(id).
+ */
+export function MortgageContextProvider({
+  value,
+  children,
+}: {
+  value: MortgageContextType;
+  children: React.ReactNode;
+}) {
   return (
     <MortgageContext.Provider value={value}>
       {children}
@@ -91,7 +26,7 @@ export function MortgageProvider({ children }: { children: React.ReactNode }) {
 export function useMortgage() {
   const context = useContext(MortgageContext);
   if (!context) {
-    throw new Error('useMortgage must be used within a MortgageProvider');
+    throw new Error('useMortgage must be used within a MortgageContextProvider (on the result page)');
   }
   return context;
 }
