--- src/services/mortgage/ServerMortgageService.ts
@@ -1,98 +1,0 @@
-import {
-  IMortgageService,
-  BaseCalculationParams,
-  MortgageCalculationParams,
-  MortgageCalculationResults,
-  AmortizationScheduleResults
-} from './IMortgageService';
-
-/**
- * Server implementation of the mortgage service
- * This implementation communicates with a server API for calculations
- * Currently this is a placeholder for future implementation
- */
-export class ServerMortgageService implements IMortgageService {
-  // This field is used in the commented-out code examples
-  // @ts-expect-error - This field is used in the commented-out code examples
-  private readonly _apiBaseUrl: string;
-
-  /**
-   * Create a new server mortgage service
-   * @param apiBaseUrl The base URL for the mortgage API
-   */
-  constructor(apiBaseUrl?: string) {
-    this._apiBaseUrl = apiBaseUrl || '/api/mortgage';
-  }
-
-  /**
-   * Base calculation (loan only). Placeholder for future server API.
-   */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async calculateBase(_params: BaseCalculationParams): Promise<MortgageCalculationResults> {
-    throw new Error('Server implementation not available yet');
-  }
-
-  /**
-   * Calculate mortgage results based on input parameters
-   * @param _params Mortgage calculation parameters
-   * @returns Mortgage calculation results
-   */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async calculateMortgage(_params: MortgageCalculationParams): Promise<MortgageCalculationResults> {
-    try {
-      // In a real implementation, this would make an API call to the server
-      // For now, we'll throw an error to indicate this is not implemented yet
-      throw new Error('Server implementation not available yet');
-      
-      // Example of how this would be implemented:
-      // const response = await fetch(`${this._apiBaseUrl}/calculate`, {
-      //   method: 'POST',
-      //   headers: {
-      //     'Content-Type': 'application/json',
-      //   },
-      //   body: JSON.stringify(_params),
-      // });
-      // 
-      // if (!response.ok) {
-      //   throw new Error(`API error: ${response.status}`);
-      // }
-      // 
-      // return await response.json();
-    } catch (error) {
-      console.error('Error calculating mortgage:', error);
-      throw error;
-    }
-  }
-
-  /**
-   * Generate an amortization schedule for a loan
-   * @param _params Mortgage calculation parameters
-   * @returns Amortization schedule results
-   */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async generateAmortizationSchedule(_params: MortgageCalculationParams): Promise<AmortizationScheduleResults> {
-    try {
-      // In a real implementation, this would make an API call to the server
-      // For now, we'll throw an error to indicate this is not implemented yet
-      throw new Error('Server implementation not available yet');
-      
-      // Example of how this would be implemented:
-      // const response = await fetch(`${this._apiBaseUrl}/amortization`, {
-      //   method: 'POST',
-      //   headers: {
-      //     'Content-Type': 'application/json',
-      //   },
-      //   body: JSON.stringify(_params),
-      // });
-      // 
-      // if (!response.ok) {
-      //   throw new Error(`API error: ${response.status}`);
-      // }
-      // 
-      // return await response.json();
-    } catch (error) {
-      console.error('Error generating amortization schedule:', error);
-      throw error;
-    }
-  }
-}
